requirements.txt
openai
portkey-ai
python-dotenv
click
paramiko
PyGithub
supabase
requests
pydantic
python-digitalocean
